time:follow POSIX use long to define time_t
MIRTOS-883
https: //man7.org/linux/man-pages/man3/clock_settime.3.html

+++ include/time.h
@@ -102,7 +102,7 @@
 
 /* Scalar types */
 
-typedef uint32_t  time_t;         /* Holds time in seconds */
+typedef long      time_t;         /* Holds time in seconds */
 typedef uint8_t   clockid_t;      /* Identifies one time base source */
 typedef FAR void *timer_t;        /* Represents one POSIX timer */
